Improve menu-item selection bounds
This primarily involved moving the menu item top and bottom padding from
the __item <div> into the salian <a> element. It reqired adding a
padding-top to the __icon-container and __menu-toggle, as well as adding
a new __item__right-border to the markup. There are some quirks yet, so
be aware of the comments added.